# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handlers.AlertsResponse:
    properties:
      alert_count:
        example: 2
        type: integer
      alerts:
        items:
          $ref: '#/definitions/radio-backend_internal_domain.Alert'
        type: array
      has_critical:
        example: false
        type: boolean
      has_warning:
        example: true
        type: boolean
    type: object
...
  internal_handlers.CheckPartitionsResponse:
    properties:
      has_missing:
        example: false
        type: boolean
      months_ahead:
        example: 3
        type: integer
      results:
        items:
          $ref: '#/definitions/radio-backend_internal_domain.PartitionCheckResult'
        type: array
    type: object
  internal_handlers.CleanupPartitionsResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/radio-backend_internal_domain.PartitionCleanupResult'
        type: array
      retention_months:
        example: 12
        type: integer
      total_dropped:
        example: 3
        type: integer
    type: object
...
  internal_handlers.ErrorDetail:
    properties:
      code:
        example: STATION_NOT_FOUND
        type: string
      field:
        example: email (only present on validation errors)
        type: string
      message:
        example: station not found
        type: string
    type: object
  internal_handlers.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/internal_handlers.ErrorDetail'
    type: object
...
  internal_handlers.PartitionStatusResponse:
    properties:
      partitions:
        items:
          $ref: '#/definitions/radio-backend_internal_domain.PartitionStatusResult'
        type: array
      total:
        example: 18
        type: integer
      total_rows:
        example: 154200
        type: integer
      total_size_mb:
        example: 42.5
        type: number
    type: object
...
  internal_handlers.RecommendationsResponse:
    properties:
      recommendations:
        items:
          $ref: '#/definitions/radio-backend_internal_domain.MaintenanceRecommendation'
        type: array
      total:
        example: 3
        type: integer
    type: object
...
  internal_handlers.RefreshStatisticsResponse:
    properties:
      days_back:
        example: 7
        type: integer
      statistics:
        items:
          $ref: '#/definitions/radio-backend_internal_domain.RefreshStatistics'
        type: array
    type: object
...
  internal_handlers.RefreshViewsResponse:
    properties:
      results:
        items:
          $ref: '#/definitions/radio-backend_internal_domain.RefreshResult'
        type: array
      type:
        example: all
        type: string
    type: object
...
  internal_handlers.SimpleErrorResponse:
    properties:
      error:
        example: authentication required
        type: string
    type: object
...
  radio-backend_internal_domain.Alert:
    properties:
      action:
        type: string
      category:
        description: database, redis, api, partitions, views, performance
        type: string
      level:
        description: info, warning, critical
        type: string
      message:
        type: string
      resolved:
        type: boolean
      timestamp:
        type: string
    type: object
...
  radio-backend_internal_domain.MaintenanceOperation:
    enum:
    - refresh_views
    - cleanup_partitions
    - check_partitions
    - partition_status
    - full_maintenance
    type: string
    x-enum-varnames:
    - MaintenanceRefreshViews
    - MaintenanceCleanupPartitions
    - MaintenanceCheckPartitions
    - MaintenancePartitionStatus
    - MaintenanceFullMaintenance
  radio-backend_internal_domain.MaintenanceRecommendation:
    properties:
      last_run_at:
        type: string
      next_run_at:
        type: string
      operation:
        $ref: '#/definitions/radio-backend_internal_domain.MaintenanceOperation'
      priority:
        description: '"critical", "warning", "info"'
        type: string
      reason:
        type: string
      should_run:
        type: boolean
    type: object
...
  radio-backend_internal_domain.PartitionCheckResult:
    properties:
      message:
        type: string
      month_date:
        type: string
      partition_name:
        type: string
      partitions_exist:
        type: boolean
      table_name:
        type: string
    type: object
  radio-backend_internal_domain.PartitionCleanupResult:
    properties:
      dropped:
        type: boolean
      message:
        type: string
      partition_date:
        type: string
      partition_name:
        type: string
      table_name:
        type: string
    type: object
...
  radio-backend_internal_domain.PartitionStatusResult:
    properties:
      index_size_mb:
        type: number
      partition_date:
        type: string
      partition_name:
        type: string
      row_count:
        type: integer
      size_mb:
        type: number
      table_name:
        type: string
      total_size_mb:
        type: number
    type: object
...
  radio-backend_internal_domain.RefreshResult:
    properties:
      completed_at:
        type: string
      duration_ms:
        type: integer
      error_message:
        type: string
      rows_affected:
        type: integer
      started_at:
        type: string
      status:
        type: string
      view_name:
        type: string
    type: object
  radio-backend_internal_domain.RefreshStatistics:
    properties:
      avg_duration_ms:
        type: number
      failed_refreshes:
        type: integer
      last_refresh:
        type: string
      last_status:
        type: string
      max_duration_ms:
        type: integer
      min_duration_ms:
        type: integer
      successful_refreshes:
        type: integer
      total_refreshes:
        type: integer
      view_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@radiobackend.com
    name: API Support
  description: "Radio streaming API with JWT authentication, analytics and multi-language
    support (i18n).\n\n**Security Features (v2.1):**\n- \U0001F510 Timing attack prevention
    (constant-time operations)\n- \U0001F512 HTTPS enforcement (production)\n- \U0001F6AB
    Account lockout (10 attempts = 30 min lockout)\n- \U0001F4E7 Email rate limiting
    (10 attempts/hour per email)\n- \U0001F511 Password strength enforcement (special
    chars + 47 common passwords blocked)\n- \U0001F6E1️ Session hijacking protection
    (User-Agent validation)\n- \U0001F4DD Security event logging (failed attempts
    with metadata)\n- ⚡ Rate limiting (5 attempts/15min per IP)\n- \U0001F510 JWT
    with RS256 + token revocation\n- \U0001F916 Bot protection via shared secret (X-Rradio-Secret
    header)\n\n**Internationalization:**\nSupported languages: Spanish (es), English
    (en), French (fr), German (de).\nUse the 'lang' parameter or 'Accept-Language'
    header to specify the desired language.\n\n**Security Score:** 100/100 (audited
    Dec 2025)"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Radio Backend API
  version: "2.1"
paths:
  /admin/maintenance/check-partitions:
    get:
      consumes:
      - application/json
      description: Verifies that partitions exist for future months to prevent insertion
        failures.
      parameters:
      - default: 3
        description: Number of months ahead to check
        in: query
        maximum: 12
        minimum: 1
        name: months_ahead
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Check results
          schema:
            $ref: '#/definitions/internal_handlers.CheckPartitionsResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Check future partitions
      tags:
      - Maintenance
  /admin/maintenance/cleanup-partitions:
    post:
      consumes:
      - application/json
      description: Removes partitions older than specified retention period to free
        up disk space.
      parameters:
      - default: 12
        description: Number of months to retain
        in: query
        maximum: 36
        minimum: 1
        name: retention_months
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cleanup results
          schema:
            $ref: '#/definitions/internal_handlers.CleanupPartitionsResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cleanup old partitions
      tags:
      - Maintenance
  /admin/maintenance/full:
    post:
      consumes:
      - application/json
      description: Executes a complete maintenance routine including view refresh,
        partition check, and status report.
      produces:
      - application/json
      responses:
        "200":
          description: Full maintenance results (dynamic map)
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Perform full maintenance
      tags:
      - Maintenance
  /admin/maintenance/partition-status:
    get:
      consumes:
      - application/json
      description: Returns detailed status of all partitions including size, row count,
        and date ranges.
      produces:
      - application/json
      responses:
        "200":
          description: Partition status
          schema:
            $ref: '#/definitions/internal_handlers.PartitionStatusResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get partition status
      tags:
      - Maintenance
  /admin/maintenance/recommendations:
    get:
      consumes:
      - application/json
      description: Returns recommendations for database maintenance operations that
        should be executed. Shows priority levels and reasons.
      produces:
      - application/json
      responses:
        "200":
          description: Maintenance recommendations
          schema:
            $ref: '#/definitions/internal_handlers.RecommendationsResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get maintenance recommendations
      tags:
      - Maintenance
  /admin/maintenance/refresh-stats:
    get:
      consumes:
      - application/json
      description: Returns statistics about materialized view refreshes including
        duration, success rate, and last refresh time.
      parameters:
      - default: 7
        description: Number of days to look back
        in: query
        maximum: 90
        minimum: 1
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Refresh statistics
          schema:
            $ref: '#/definitions/internal_handlers.RefreshStatisticsResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get refresh statistics
      tags:
      - Maintenance
  /admin/maintenance/refresh-views:
    post:
      consumes:
      - application/json
      description: Refreshes materialized views used for SEO and analytics. Can refresh
        all views or specific types (seo/analytics).
      parameters:
      - default: all
        description: 'Type of views to refresh: all, seo, analytics'
        enum:
        - all
        - seo
        - analytics
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Refresh results
          schema:
            $ref: '#/definitions/internal_handlers.RefreshViewsResponse'
        "401":
          description: Invalid or missing authentication token
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Access denied - Admin users only
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Refresh materialized views
      tags:
      - Maintenance
  /admin/monitoring/alerts:
    get:
      description: |-
        Returns list of active system alerts based on health checks
        Alerts include partition issues, stale materialized views, and other problems
      produces:
      - application/json
      responses:
        "200":
          description: Alerts retrieved successfully
          schema:
            $ref: '#/definitions/internal_handlers.AlertsResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "403":
          description: Admin access required
          schema:
            $ref: '#/definitions/internal_handlers.SimpleErrorResponse'
        "500":
          description: Failed to get alerts
          schema:
            $ref: '#/definitions/internal_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get system alerts
      tags:
      - Monitoring
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  SharedSecret:
    description: Shared secret key required for public station and SEO endpoints (bot
      protection).
    in: header
    name: X-Rradio-Secret
    type: apiKey
swagger: "2.0"